notification update

+++ server/src/main/java/com/example/prep_pilot/service/FollowsService.java
@@ -1,12 +1,16 @@
 package com.example.prep_pilot.service;
 
 import com.example.prep_pilot.dto.FollowsDto;
+import com.example.prep_pilot.dto.NotificationDto;
 import com.example.prep_pilot.entity.Follows;
+import com.example.prep_pilot.entity.Notifications;
 import com.example.prep_pilot.entity.User;
 import com.example.prep_pilot.exception.UserNotFoundException;
 import com.example.prep_pilot.repository.FollowsRepository;
+import com.example.prep_pilot.repository.NotificationRepository;
 import com.example.prep_pilot.repository.UserRepository;
 import jakarta.transaction.Transactional;
+import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -19,11 +23,15 @@ public class FollowsService {
 
     private final FollowsRepository followsRepository;
     private final UserRepository userRepository;
+    private final NotificationRepository notificationRepository;
+    private final SimpMessagingTemplate messagingTemplate;
 
-    public FollowsService(FollowsRepository followsRepository, UserRepository userRepository){
+    public FollowsService(FollowsRepository followsRepository, UserRepository userRepository, NotificationRepository notificationRepository, SimpMessagingTemplate messagingTemplate){
 
         this.followsRepository = followsRepository;
         this.userRepository = userRepository;
+        this.notificationRepository = notificationRepository;
+        this.messagingTemplate = messagingTemplate;
     }
 
     @Transactional
@@ -33,11 +41,19 @@ public class FollowsService {
         User followed = userRepository.findById(userId).orElseThrow(() ->
                 new UserNotFoundException(userId)
         );
+        if(following.equals(followed))
+            return;
         Optional<Follows> follows = followsRepository.findByFollowerUserIdAndFollowingUserUsername(userId, username);
         if(follows.isPresent())
             followsRepository.delete(follows.get());
-        else
+        else {
             followsRepository.save(new Follows(null, followed, following, LocalDateTime.now()));
+            Notifications notification = Notifications.createFollowNotification(following, followed);
+            notificationRepository.save(notification);
+            NotificationDto notificationDto = NotificationDto.toDto(notification);
+            // "/topic/notifications/{userId}" 구독한 사용자에게 메시지 전송
+            messagingTemplate.convertAndSend("/topic/notifications/" + followed.getId(), notificationDto);
+        }
     }
 
     public Boolean isFollowing(String username, Long userId) {
